+++ src/module/vue/components/sf-character-sheet-tabs/sf-connections.vue
@@ -1,13 +1,20 @@
 <template>
   <div class="flexcol">
     <transition-group name="slide" tag="div" class="nogrow">
-      <progress-box
-        v-for="item in connections"
+      <div
+        class="flexrow nogrow"
+        v-for="(item, i) in connections"
         :key="item._id"
-        :item="item"
-        :actor="actor"
-        :showStar="true"
-      />
+      >
+        <order-buttons
+          v-if="editMode"
+          :i="i"
+          :length="connections.length"
+          @sortUp="sortUp"
+          @sortDown="sortDown"
+        />
+        <progress-box :item="item" :actor="actor" :showStar="true" />
+      </div>
     </transition-group>
 
     <div class="flexrow nogrow" style="text-align: center">
@@ -37,6 +44,11 @@ export default {
       return this.actor.items
         .filter((x) => x.type === 'progress')
         .filter((x) => x.data.subtype === 'bond') // legacy name
+        .sort((a, b) => (a.sort || 0) - (b.sort || 0))
+    },
+
+    editMode() {
+      return this.actor.flags['foundry-ironsworn']?.['edit-mode']
     },
   },
 
@@ -52,6 +64,29 @@ export default {
         { parent: this.$actor }
       )
       item.sheet.render(true)
+    },
+
+    async applySort(oldI, newI, sortBefore) {
+      const foundryItems = this.$actor.items
+        .filter((x) => x.type === 'progress')
+        .filter((x) => x.data.data.subtype === 'bond')
+        .sort((a, b) => (a.data.sort || 0) - (b.data.sort || 0))
+
+      const updates = SortingHelpers.performIntegerSort(foundryItems[oldI], {
+        target: foundryItems[newI],
+        siblings: foundryItems,
+        sortBefore,
+      })
+      await Promise.all(
+        updates.map(({ target, update }) => target.update(update))
+      )
+    },
+
+    sortUp(i) {
+      this.applySort(i, i - 1, true)
+    },
+    sortDown(i) {
+      this.applySort(i, i + 1, false)
     },
   },
 }