+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Re-ordering for assets ([#368](https://github.com/ben/foundry-ironsworn/pull/368))
 - Re-ordering for progresses ([#369](https://github.com/ben/foundry-ironsworn/pull/369))
+- Re-ordering for connections
 
 ## 1.15.2